requirements.txt
Django==1.11.16
django-modeltranslation==0.12.2
django-seo2==1.0.1
django-wysiwyg-redactor==0.5.1
olefile==0.44
Pillow==4.2.1
pytz==2018.7
six==1.11.0
